+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # 0.1.0 (Unreleased)
 
 - Define the available types of packet in the `PacketType` enum.
+- Define the `Packet` class as the super-type of all packets.